Remove the now unused DB modules: load, tables, queries
