+++ portingdb/cli.py
@@ -6,7 +6,6 @@ from pathlib import Path
 
 import click
 
-from portingdb import tables
 from portingdb.load_data import get_data
 from portingdb.check_drops import check_drops
